moves training args parser to argparsing.py
